+++ project/src/components/play-button/play-button.tsx
@@ -1,5 +1,5 @@
 import {Link} from 'react-router-dom';
-import {PlayButtonIconAndText} from '../play-pause-button/play_button';
+import {PlayButtonIconAndText} from '../play-pause-button/play-button-icon-and-text';
 import {PLAYER_ROUTE} from '../../constants';
 
 class PlayButtonProps {
