+++ project/src/components/play-pause-button/pause-button-icon-and-text.tsx
